refactor: local storage util 파일명 변경
